Rename tower-middleware to tower-layer and ext feature to util
