+++ Cargo.toml
@@ -32,7 +32,7 @@ members = [
   "tower-filter",
   "tower-in-flight-limit",
   "tower-mock",
-  "tower-middleware",
+  "tower-layer",
   "tower-rate-limit",
   "tower-reconnect",
   "tower-retry",
