+++ tower-in-flight-limit/Cargo.toml
@@ -7,7 +7,7 @@ publish = false
 [dependencies]
 futures = "0.1"
 tower-service = { version = "0.2", path = "../tower-service" }
-tower-middleware = { version = "0.1", path = "../tower-middleware" }
+tower-layer = { version = "0.1", path = "../tower-layer" }
 
 [dev-dependencies]
 tokio-test = { git = "https://github.com/carllerche/tokio-test" }
